Renamed `Multi*Based` classes to `MultipleChecksOnA*` for clarity
Previous naming described implementation rather than intended behavior
