+++ test/unit/Comparator/BackwardsCompatibility/ClassBased/MultipleChecksOnAClassTest.php
@@ -2,28 +2,31 @@
 
 declare(strict_types=1);
 
-namespace RoaveTest\ApiCompare\Comparator\BackwardsCompatibility\InterfaceBased;
+namespace RoaveTest\ApiCompare\Comparator\BackwardsCompatibility\ClassBased;
 
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Roave\ApiCompare\Change;
 use Roave\ApiCompare\Changes;
-use Roave\ApiCompare\Comparator\BackwardsCompatibility\InterfaceBased\InterfaceBased;
-use Roave\ApiCompare\Comparator\BackwardsCompatibility\InterfaceBased\MultiInterfaceBased;
+use Roave\ApiCompare\Comparator\BackwardsCompatibility\ClassBased\ClassBased;
+use Roave\ApiCompare\Comparator\BackwardsCompatibility\ClassBased\MultipleChecksOnAClass;
 use Roave\BetterReflection\Reflection\ReflectionClass;
 
-final class MultiInterfaceBasedTest extends TestCase
+/**
+ * @covers \Roave\ApiCompare\Comparator\BackwardsCompatibility\ClassBased\MultipleChecksOnAClass
+ */
+final class MultipleChecksOnAClassTest extends TestCase
 {
     public function testChecksAllGivenCheckers() : void
     {
-        /** @var InterfaceBased|MockObject $checker1 */
-        $checker1 = $this->createMock(InterfaceBased::class);
-        /** @var InterfaceBased|MockObject $checker2 */
-        $checker2 = $this->createMock(InterfaceBased::class);
-        /** @var InterfaceBased|MockObject $checker3 */
-        $checker3 = $this->createMock(InterfaceBased::class);
+        /** @var ClassBased|MockObject $checker1 */
+        $checker1 = $this->createMock(ClassBased::class);
+        /** @var ClassBased|MockObject $checker2 */
+        $checker2 = $this->createMock(ClassBased::class);
+        /** @var ClassBased|MockObject $checker3 */
+        $checker3 = $this->createMock(ClassBased::class);
 
-        $multiCheck = new MultiInterfaceBased($checker1, $checker2, $checker3);
+        $multiCheck = new MultipleChecksOnAClass($checker1, $checker2, $checker3);
 
         /** @var ReflectionClass|MockObject $from */
         $from = $this->createMock(ReflectionClass::class);
